♻️ refactor: ディレクトリ名の変更 #9
