+++ pages/contact.jsx
@@ -1,8 +1,8 @@
 // contactページ
 
-import Meta from '@/src/components/Elements/meta/Meta'
+import Meta from '@/src/components/Elements/Meta/Meta'
 import Container from '@/src/components/Layouts/Container/Container'
-import Hero from '@/src/components/Elements/hero/Hero'
+import Hero from '@/src/components/Elements/Hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/contact.jpg'
 import Form from '@/src/components/Form/form'
