+++ src/components/layouts/footer/Footer.jsx
@@ -1,6 +1,6 @@
 import styles from './Footer.module.css'
-import Social from '../../Elements/social/Social'
-import Logo from '../../Elements/logo/logo'
+import Social from '../../Elements/Social/Social'
+import Logo from '../../Elements/Logo/logo'
 
 export default function Footer() {
   return (
